pyproject.toml
[project]
name = "event-booking"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.17.2",
    "argon2-cffi>=25.1.0",
    "asgiref>=3.11.0",
    "asyncpg>=0.31.0",
    "email-validator>=2.3.0",
    "fastapi>=0.123.5",
    "fastapi-mail>=1.5.8",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.12.0",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "redis>=7.1.0",
    "requests>=2.32.5",
    "slowapi>=0.1.9",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.38.0",
]

[dependency-groups]
dev = [
    "aiosqlite>=0.21.0",
    "httpx>=0.28.1",
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
]
